src/africa/semicolon/ecommerce/Order.java
package africa.semicolon.ecommerce;

public class Order {
    private int orderId;
    private String dateOrdered;
    private String dateShipped;
    private String customerName;
    private String customerAddress;
    private String shippingDetails;

    public void setOrderId(int orderId) {
        this.orderId = orderId;
    }

    public int getOrderId() {
        return orderId;
    }

    public void setDateOrdered(String dateOrdered) {
        this.dateOrdered = dateOrdered;
    }

    public String getDateOrdered() {
        return dateOrdered;
    }

    public void setDateShipped(String dateShipped) {
        this.dateShipped = dateShipped;
    }

    public String getDateShipped() {
        return dateShipped;
    }

    public void setCustomerName(String customerName) {
        this.customerName = customerName;
    }

    public String getCustomerName() {
        return customerName;
    }

    public void setCustomerAddress(String customerAddress) {
        this.customerAddress = customerAddress;
    }

    public String getCustomerAddress() {
        return customerAddress;
    }

    public void setShippingDetails(String shippingDetails) {
        this.shippingDetails = shippingDetails;
    }

    public String getShippingDetails() {
        return shippingDetails;
    }
}
